requirements.txt
httpx==0.28.1
pydantic==2.11.1
python-dotenv==1.0.1
openai==1.68.2
anthropic==0.49.0
google-generativeai==0.8.4
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
